+++ pkg/log/wrapper.go
@@ -1,0 +1,9 @@
+package log
+
+type Logger interface {
+	Debug(msg string)
+	Info(msg string)
+	Warn(msg string)
+	Error(msg string)
+	Fatal(msg string)
+}